Project and namespace reorganization
