+++ src/Auth0.Core/Exceptions/ApiException.cs
@@ -2,11 +2,7 @@
 using System.Net;
 using Auth0.Core;
 
-#if MANAGEMENT_API
-namespace Auth0.ManagementApi.Client.Exceptions
-#elif AUTHENTICATION_API
-namespace Auth0.AuthenticationApi.Client.Exceptions
-#endif
+namespace Auth0.Core.Exceptions
 {
     /// <summary>
     /// Represents errors that occur when making API calls.
