+++ src/Auth0.ManagementApi.Client/Clients/ConnectionsClient.cs
@@ -3,6 +3,7 @@ using Auth0.ManagementApi.Client.Models;
 using PortableRest;
 using System.Collections.Generic;
 using System.Threading.Tasks;
+using Auth0.Core.Http;
 
 namespace Auth0.ManagementApi.Client.Clients
 {
