+++ src/Auth0.ManagementApi.Client/Clients/JobsClient.cs
@@ -2,9 +2,9 @@
 using System.IO;
 using System.Threading.Tasks;
 using Auth0.Core;
+using Auth0.Core.Http;
 using Auth0.ManagementApi.Client.Models;
 using PortableRest;
-using Auth0.Core.ApiClients;
 
 namespace Auth0.ManagementApi.Client.Clients
 {
