+++ src/Auth0.ManagementApi.Client/Clients/StatsClient.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 using Auth0.Core;
+using Auth0.Core.Http;
 
 namespace Auth0.ManagementApi.Client.Clients
 {
