+++ src/Auth0.ManagementApi.Client/Clients/TicketsClient.cs
@@ -1,5 +1,6 @@
 ﻿using System.Threading.Tasks;
 using Auth0.Core;
+using Auth0.Core.Http;
 using PortableRest;
 using Auth0.ManagementApi.Client.Models;
