+++ src/Auth0.ManagementApi.Client/Clients/UsersClient.cs
@@ -1,9 +1,9 @@
 ﻿using System.Collections.Generic;
 using System.Threading.Tasks;
 using Auth0.Core;
+using Auth0.Core.Http;
 using PortableRest;
 using Auth0.ManagementApi.Client.Models;
-using Auth0.Core.ApiClients;
 
 namespace Auth0.ManagementApi.Client.Clients
 {
