+++ src/Auth0.ManagementApi.Client/Core/Http/DiagnosticsHeader.cs
@@ -6,11 +6,8 @@ using System.Text;
 using System.Xml.Linq;
 using Newtonsoft.Json;
 
-#if MANAGEMENT_API
-namespace Auth0.ManagementApi.Client.Diagnostics
-#elif AUTHENTICATION_API
-namespace Auth0.AuthenticationApi.Client.Diagnostics
-#endif
+// ReSharper disable once CheckNamespace
+namespace Auth0.Core.Http
 {
     /// <summary>
     ///     Represents important information pertaining to the SDK that is sent to Auth0 for diagnostic purposes.
