+++ src/Auth0.ManagementApi.Client/Core/Http/IApiConnection.cs
@@ -1,19 +1,9 @@
 ﻿using System.Collections.Generic;
 using System.Threading.Tasks;
 using PortableRest;
-using Auth0.Core.ApiClients;
 
-#if MANAGEMENT_API
-using Auth0.ManagementApi.Client.Models;
-#elif AUTHENTICATION_API
-using Auth0.AuthenticationApi.Client.Models;
-#endif
-
-#if MANAGEMENT_API
-namespace Auth0.ManagementApi.Client
-#elif AUTHENTICATION_API
-namespace Auth0.AuthenticationApi.Client
-#endif
+// ReSharper disable once CheckNamespace
+namespace Auth0.Core.Http
 {
     /// <summary>
     /// The communication layer between the various API clients and the actual API backend. All API calls happen through this interface.
