+++ tests/Auth0.ManagementApi.Client.IntegrationTests/EmailProviderTests.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Threading.Tasks;
 using Auth0.Core;
-using Auth0.ManagementApi.Client.Exceptions;
+using Auth0.Core.Exceptions;
 using FluentAssertions;
 using NUnit.Framework;
 using Auth0.ManagementApi.Client.Models;
